webrestaurante/requirements.txt
Django==2.2.3
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework==3.9.1
Markdown==3.0.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
sqlparse==0.4.1
